requirements.txt
coverage==4.5.1
Django==1.11.15
django-rest-framework==0.1.0
djangorestframework==3.8.2
psycopg2-binary==2.7.5
pytz==2018.5
rest-condition==1.0.3
